pyproject.toml
[tool.poetry]
name = "embeddings"
version = "0.1.0"
description = ""
authors = ["daragh <dmoran@eisgroup.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
tiktoken = "^0.6.0"
langchain = "^0.1.17"
openai = "^1.25.1"
numpy = "^1.26.4"
scipy = "^1.13.0"
python-telegram-bot = "^21.1.1"
python-dotenv = "^1.0.1"
cairosvg = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
